Change int to float

+++ HWPLDS-2/FLOWRATE1.h
@@ -10,7 +10,7 @@ unsigned long oldTime;
 
 class Flowrate1{
   public:
-    int flr1_loop();
+    float flr1_loop();
     static void pulseCounter();
     void setup1();
   };
@@ -33,10 +33,10 @@ void Flowrate1::setup1()
   attachInterrupt(sensorInterrupt, pulseCounter, FALLING);
 }
 
-int Flowrate1::flr1_loop()
+float Flowrate1::flr1_loop()
 {
    
-   if((millis() - oldTime) > 1000)    // Only process counters once per second
+   if((millis() - oldTime) > 1000)                // Only process counters once per second
   { 
     detachInterrupt(sensorInterrupt);
 
@@ -46,15 +46,15 @@ int Flowrate1::flr1_loop()
     totalMilliLitres += flowMilliLitres;
       
       unsigned int frac;
-//      Serial.print(int(flowRate));  // Print the integer part of the variable
-//      Serial.print("L/min ");       
+//    Serial.print(int(flowRate));                // Print the integer part of the variable
+//    Serial.print("L/min ");       
 //    Serial.print(totalMilliLitres);
 //    Serial.println("mL"); 
  
     pulseCount = 0;
  
     attachInterrupt(sensorInterrupt, pulseCounter, FALLING);
-    return int(flowRate);
+    return flowRate;
   }
      
 }
