+++ HWPLDS-2/FLOWRATE2.h
@@ -11,7 +11,7 @@ unsigned long oldTime1;
 
 class Flowrate2{
   public:
-    int flr2_loop();
+    float flr2_loop();
     static void pulseCount1er();
     void setup2();
   };
@@ -36,7 +36,7 @@ void Flowrate2::setup2()
   attachInterrupt(sensorInterrupt1, pulseCount1er, FALLING);
 }
 
-int Flowrate2::flr2_loop()
+float Flowrate2::flr2_loop()
 {
    
    if((millis() - oldTime1) > 1000)    // Only process counters once per second
@@ -62,7 +62,7 @@ int Flowrate2::flr2_loop()
     
     // Enable the interrupt again now that we've finished sending output
     attachInterrupt(sensorInterrupt1, pulseCount1er, FALLING);
-    return int(flowRate1);
+    return flowRate1;
   }
  
 }
